requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
PyMuPDF